requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.2
colorama==0.4.6
crispy-bootstrap4==2024.10
Django==5.1.6
django-bootstrap4==24.4
django-crispy-forms==2.3
django-formtools==2.5.1
django-lifecycle==1.2.4
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-safedelete==1.4.0
django-simple-history==3.8.0
django-two-factor-auth==1.17.0
packaging==24.2
phonenumbers==8.13.54
pillow==11.1.0
pip==25.0
psycopg2-binary==2.9.10
pymongo==3.11.4
pyotp==2.9.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2025.1
qrcode==7.4.2
requests==2.31.0
cinetpay-sdk==1.0.0
setuptools==75.8.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
